documentor/storage/__init__.py
"""
Vector storage implementations.
"""

from documetor.storage.base import VectorStore
from documetor.storage.local import LocalVectorStore
from documetor.storage.vertex import VertexMatchingEngineStore

__all__ = ['VectorStore', 'LocalVectorStore', 'VertexMatchingEngineStore']
